Usage of heatmap for the website
This commit creates a new website which uses heat map to help track to
where the user have seen the website. I plan to make it such that if the
user sees the left side of the screen more, then the left option is
chosen, but if the user sees the right side of the screen more, then the
right option is selected.