Halfway there to upload activity.
This commit makes the user able to upload the image. The image is checked in terms of contents and extension, whether the uploaded image is actually an image file or not. Once the image is okay, the filename is then hashed using SHA256 based on the current time. Now, the only task remaining for uploading the image is to actually upload the image to the server, and store the activity details into the database.