Add options to change the time to look over the options and the refresh.
This commit adds the options for the caregiver to change the time for the patient to look to one options, as well as time for the website to refresh itself, as requested by Doctor Nita. This allows the caregiver to adapt to different focus times for each patient. However, the time is still not yet functional.