Make the program able to change the volume, as well as the voice.
ResponsiveJS has voices for both male and female Indonesian voices; IIRC Mr. Eka asked such that we should be able to change the volume, as well as changing the voice between male and female. This is achieved by using determineVoice and determineVolume function to get the values from the voice and volume inputs, and every time these values are changed, we also reset the timer in changeSomething method.